+++ src/postcss/shared.ts
@@ -21,9 +21,3 @@ export function internalCssSelectorReplacer(
     map: escapeMap
   })
 }
-
-export const cssUnescape = (str: string) => {
-  return str.replaceAll(/\\([\dA-Fa-f]{1,6}[\t\n\f\r ]?|[\S\s])/g, (match) => {
-    return match.length > 2 ? String.fromCodePoint(Number.parseInt(match.slice(1).trim(), 16)) : match[1]
-  })
-}