+++ test/patcher.test.ts
@@ -17,6 +17,19 @@ describe('patcher unit test', () => {
     expect(res).toBe(undefined)
   })
 
+  it('found pkg', () => {
+    const res = getInstalledPkgJsonPath({
+      dangerousOptions: {
+        packageName: 'tailwindcss'
+      },
+      units: ['rpx']
+    })
+    expect(res).toBeTruthy()
+    const isStr = typeof res === 'string'
+    expect(isStr).toBe(true)
+    isStr && expect(res.length > 0).toBe(true)
+  })
+
   it('patch warning', () => {
     const patch = createPatch({
       units: ['rpx'],
@@ -29,4 +42,15 @@ describe('patcher unit test', () => {
     })
     expect(patch()).toBe(undefined)
   })
+
+  // it('patch snap', () => {
+  //   // const patch = createPatch({
+  //   //   units: ['rpx'],
+  //   //   dangerousOptions: {
+  //   //     gteVersion: '3.0.0',
+  //   //     packageName: 'tailwindcss'
+  //   //   }
+  //   // })
+  //   // expect(patch()).toMatchSnapshot()
+  // })
 })