+++ test/pref.test.ts
@@ -1,5 +1,5 @@
 import { wxmlCasePath, createGetCase } from './util'
-import { templeteHandler } from '@/wxml'
+import { templateHandler } from '@/wxml'
 
 const getCase = createGetCase(wxmlCasePath)
 describe('performance', () => {
@@ -9,7 +9,7 @@ describe('performance', () => {
   test('long time', async () => {
     const now = Date.now()
     const source = await getCase('pref.wxml')
-    const str = templeteHandler(source)
+    const str = templateHandler(source)
 
     const ts = Date.now() - now
     expect(ts < 1000).toBe(true)