+++ test/wxml/handlingWhitespace.test.ts
@@ -1,11 +1,11 @@
 /* eslint-disable no-useless-escape */
-import { templeteHandler } from '@/wxml/index'
+import { templateHandler } from '@/wxml/index'
 // import { format } from '../util'
 
 describe('handling whitespace', () => {
   test('grid grid-cols-[1fr_500px_2fr]', () => {
     expect(
-      templeteHandler(`<div class="grid grid-cols-[1fr_500px_2fr]">
+      templateHandler(`<div class="grid grid-cols-[1fr_500px_2fr]">
       <!-- ... -->
     </div>`)
     ).toMatchSnapshot()
@@ -13,7 +13,7 @@ describe('handling whitespace', () => {
 
   test("bg-[url('/what_a_rush.png')]", () => {
     expect(
-      templeteHandler(`<div class="bg-[url('/what_a_rush.png')]">
+      templateHandler(`<div class="bg-[url('/what_a_rush.png')]">
       <!-- ... -->
     </div>`)
     ).toMatchSnapshot()
@@ -21,7 +21,7 @@ describe('handling whitespace', () => {
 
   test("before:content-['hello_world']", () => {
     expect(
-      templeteHandler(`<div class="before:content-['hello\_world']">
+      templateHandler(`<div class="before:content-['hello\_world']">
       <!-- ... -->
     </div>`)
     ).toMatchSnapshot()