+++ test/wxml/resolvingAmbiguities.test.ts
@@ -1,24 +1,24 @@
-import { templeteHandler } from '@/wxml/index'
+import { templateHandler } from '@/wxml/index'
 // import { format } from '../util'
 
 describe('resolving ambiguities', () => {
   test('text-[22px]', () => {
-    expect(templeteHandler('<div class="text-[22px]">...</div>')).toMatchSnapshot()
+    expect(templateHandler('<div class="text-[22px]">...</div>')).toMatchSnapshot()
   })
 
   test('text-[#bada55]', () => {
-    expect(templeteHandler('<div class="text-[#bada55]">...</div>')).toMatchSnapshot()
+    expect(templateHandler('<div class="text-[#bada55]">...</div>')).toMatchSnapshot()
   })
 
   test('text-[var(--my-var)]', () => {
-    expect(templeteHandler('<div class="text-[var(--my-var)]">...</div>')).toMatchSnapshot()
+    expect(templateHandler('<div class="text-[var(--my-var)]">...</div>')).toMatchSnapshot()
   })
 
   test('text-[length:var(--my-var)]', () => {
-    expect(templeteHandler('<div class="text-[length:var(--my-var)]">...</div>')).toMatchSnapshot()
+    expect(templateHandler('<div class="text-[length:var(--my-var)]">...</div>')).toMatchSnapshot()
   })
 
   test('text-[color:var(--my-var)]', () => {
-    expect(templeteHandler('<div class="text-[color:var(--my-var)]">...</div>')).toMatchSnapshot()
+    expect(templateHandler('<div class="text-[color:var(--my-var)]">...</div>')).toMatchSnapshot()
   })
 })